On Windows, convert paths to use forward slashes before fnmatch call
